Centralize schema variant views
Centralize schema variant view collection into the dal. The integration
test using them now checks for count, but it may be able to test for
more in the future.
